python/monitor/monitor_diskgroup/requirements.txt
Click==7.0
PyYAML==3.13
requests==2.19.1